refactor: 시스템 메시지 상수화
