+++ src/App.js
@@ -1,6 +1,6 @@
 import RaceController from "./controllers/RaceController.js";
-import InputView from "./views/InputView.js";
-import OutputView from "./views/OutputView.js";
+import InputView from "./view/InputView.js";
+import OutputView from "./view/OutputView.js";
 
 class App {
   async run() {
