+++ src/constants/SystemMessage.js
@@ -1,0 +1,8 @@
+export const SYSTEM_MESSAGE = Object.freeze({
+  INPUT_CAR_NAME:
+    "경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).\n",
+  INPUT_TRY_COUNT: "시도할 횟수는 몇 회인가요?\n",
+  OUTPUT_RACE_HEADER: "\n실행 결과",
+  OUTPUT_CAR_MARK: "-",
+  OUTPUT_WINNER: "최종 우승자: ",
+});
