+++ src/view/InputView.js
@@ -1,14 +1,13 @@
+import { SYSTEM_MESSAGE } from "../constants/SystemMessage.js";
 import readLineAsync from "../utils/readLineAsync.js";
 
 class InputView {
   async getCarNames() {
-    return await readLineAsync(
-      "경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).\n",
-    );
+    return await readLineAsync(SYSTEM_MESSAGE.INPUT_CAR_NAME);
   }
 
   async getTryCount() {
-    return await readLineAsync("시도할 횟수는 몇 회인가요?\n");
+    return await readLineAsync(SYSTEM_MESSAGE.INPUT_TRY_COUNT);
   }
 }
