+++ src/view/OutputView.js
@@ -1,17 +1,21 @@
+import { SYSTEM_MESSAGE } from "../constants/SystemMessage.js";
+
 class OutputView {
   printResultHeader() {
-    console.log("실행 결과");
+    console.log(SYSTEM_MESSAGE.OUTPUT_RACE_HEADER);
   }
 
   printRaceResult(carList) {
     // UI를 위한 get은 허용한다.
     carList.forEach((car) => {
-      console.log(`${car.name} : ${"-".repeat(car.position)}`);
+      console.log(
+        `${car.name} : ${SYSTEM_MESSAGE.OUTPUT_CAR_MARK.repeat(car.position)}`,
+      );
     });
   }
 
   printWinners(winners) {
-    console.log(`최종 우승자: ${winners.join(", ")}`);
+    console.log(`${SYSTEM_MESSAGE.OUTPUT_WINNER}${winners.join(", ")}`);
   }
 
   printNewLine() {
